Added selection support for primitive values
